# docs/api/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "description" : "Jornada de trabalho com marcos de hidratação e exercício.",
    "title" : "Pausa Ativa API",
    "version" : "0.3.0"
  },
  "servers" : [ {
    "url" : "/",
    "description" : "Mesma origem do frontend"
  } ],
...
  "paths" : {
...
    "/api/v1/marcos/{id}/adiamento" : {
      "post" : {
        "description" : "O bloco seguinte passa a ter 10 min e decide o destino dos dois. Só vale uma vez por cadeia e nunca no último bloco do dia; a água não é adiada.",
        "operationId" : "adiar",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Jornada"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Bad Request"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Conflict"
          }
        },
        "summary" : "Adia o bloco de exercício. Repetir devolve a mesma situação.",
        "tags" : [ "Marco" ]
      }
    },
    "/api/v1/marcos/{id}/conclusao" : {
      "post" : {
        "operationId" : "concluir",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Jornada"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Bad Request"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Conflict"
          }
        },
        "summary" : "Conclui o marco. Repetir devolve a mesma situação.",
        "tags" : [ "Marco" ]
      }
    },
    "/api/v1/marcos/{id}/correcao" : {
      "post" : {
        "description" : "Concluído vira falha, ou o contrário, até o fim do dia em que a jornada começou. Num par adiado, corrige os dois. O lembrete fica marcado como editado.",
        "operationId" : "corrigir",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Correcao"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Jornada"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Bad Request"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Conflict"
          }
        },
        "summary" : "Corrige a resposta de um lembrete de hoje. Corrigir para a situação atual devolve a mesma.",
        "tags" : [ "Marco" ]
      }
    },
    "/api/v1/marcos/{id}/falha" : {
      "post" : {
        "operationId" : "falhar",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Jornada"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Bad Request"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Conflict"
          }
        },
        "summary" : "Marca falha no marco. Repetir devolve a mesma situação.",
        "tags" : [ "Marco" ]
      }
    },
    "/api/v1/marcos/{id}/recebimento" : {
      "post" : {
        "description" : "Sem a confirmação, o prazo vencido vira NAO_ENTREGUE em vez de FALHA.",
        "operationId" : "confirmarRecebimento",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Bad Request"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Conflict"
          }
        },
        "summary" : "Confirma que o marco chegou à tela",
        "tags" : [ "Marco" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "Bloco" : {
        "type" : "object",
        "description" : "Bloco de exercício do marco, como foi proposto no disparo.",
        "properties" : {
          "compensaAdiamento" : {
            "type" : "boolean",
            "description" : "Se o bloco compensa o anterior, que foi adiado; nesse caso tem 10 min"
          },
          "duracaoMin" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Duração do bloco: 5 ou 10 min",
            "example" : 5
          },
          "itens" : {
            "type" : "array",
            "description" : "Exercícios na ordem em que devem ser feitos",
            "items" : {
              "$ref" : "#/components/schemas/ItemDoBloco"
            }
          },
          "segundosEstimados" : {
            "type" : "integer",
            "format" : "int64",
            "description" : "Soma das estimativas dos exercícios, com as trocas",
            "example" : 278
          }
        },
        "required" : [ "compensaAdiamento", "duracaoMin", "itens", "segundosEstimados" ]
      },
...
      "Correcao" : {
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "string",
            "description" : "A situação correta do lembrete: concluído ou falha",
            "enum" : [ "CONCLUIDO", "FALHA" ]
          }
        },
        "required" : [ "status" ]
      },
...
      "ItemDoBloco" : {
        "type" : "object",
        "properties" : {
          "exercicio" : {
            "type" : "string",
            "example" : "Sentar e levantar da cadeira"
          },
          "grupo" : {
            "type" : "string",
            "example" : "Pernas"
          },
          "instrucao" : {
            "type" : "string",
            "example" : "Sente e levante da cadeira sem usar as mãos, com os pés na largura do quadril."
          },
          "quantidade" : {
            "type" : "string",
            "description" : "\"10 repetições\", \"6 por lado\" ou \"20 s\"",
            "example" : "10 repetições"
          }
        },
        "required" : [ "exercicio", "grupo", "instrucao", "quantidade" ]
      },
      "Jornada" : {
        "type" : "object",
        "description" : "Situação da jornada no instante calculadoEm. Horários no fuso America/Sao_Paulo.",
        "properties" : {
          "aguaIngeridaMl" : {
            "type" : "number",
            "description" : "Soma dos marcos de hidratação concluídos",
            "example" : 562.5
          },
          "calculadoEm" : {
            "type" : "string",
            "format" : "date-time"
          },
          "dataReferencia" : {
            "type" : "string",
            "format" : "date",
            "description" : "Dia em que a jornada começou"
          },
          "duracaoBlocoMin" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Duração dos blocos de exercício escolhida ao iniciar o dia. O bloco que compensa um adiamento tem 10 min.",
            "example" : 5
          },
          "finalizadaEm" : {
            "type" : [ "string", "null" ],
            "format" : "date-time"
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "iniciadaEm" : {
            "type" : "string",
            "format" : "date-time"
          },
          "marcos" : {
            "type" : "array",
            "description" : "Água e exercício em ordem de horário; na hora cheia, a água vem antes",
            "items" : {
              "$ref" : "#/components/schemas/Marco"
            }
          },
          "metaAguaMl" : {
            "type" : "integer",
            "format" : "int32",
            "example" : 3000
          },
          "pausadaDesde" : {
            "type" : [ "string", "null" ],
            "format" : "date-time",
            "description" : "Início da pausa em curso; nulo fora da pausa"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "EM_ANDAMENTO", "PAUSADA", "FINALIZADA", "ENCERRADA_AUTOMATICAMENTE" ]
          },
          "tempoTrabalhadoSegundos" : {
            "type" : "integer",
            "format" : "int64",
            "description" : "Tempo trabalhado até calculadoEm, sem as pausas. Enquanto a jornada estiver em andamento, a tela soma o tempo decorrido desde então.",
            "example" : 5400
          }
        },
        "required" : [ "aguaIngeridaMl", "calculadoEm", "dataReferencia", "duracaoBlocoMin", "finalizadaEm", "id", "iniciadaEm", "marcos", "metaAguaMl", "pausadaDesde", "status", "tempoTrabalhadoSegundos" ]
      },
      "Marco" : {
        "type" : "object",
        "description" : "Lembrete da jornada. Horários no fuso America/Sao_Paulo.",
        "properties" : {
          "bloco" : {
            "description" : "Bloco de exercício. Nulo na água e no exercício que ainda não disparou.",
            "oneOf" : [ {
              "$ref" : "#/components/schemas/Bloco"
            }, {
              "type" : "null"
            } ]
          },
          "categoria" : {
            "type" : "string",
            "enum" : [ "HIDRATACAO", "EXERCICIO" ]
          },
          "disparadoEm" : {
            "type" : [ "string", "null" ],
            "format" : "date-time"
          },
          "editadoEm" : {
            "type" : [ "string", "null" ],
            "format" : "date-time",
            "description" : "Instante da última correção. Nulo se o lembrete nunca foi corrigido."
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "mensagem" : {
            "type" : "string",
            "example" : "Beba ~190 ml. Levante-se para buscar a água."
          },
          "podeAdiar" : {
            "type" : "boolean",
            "description" : "Se o botão Adiar vale agora: só no exercício pendente que não compensa um adiamento e não é o último do dia"
          },
          "podeCorrigir" : {
            "type" : "boolean",
            "description" : "Se o botão Corrigir vale agora: só no lembrete concluído ou com falha, até o fim do dia em que a jornada começou"
          },
          "recebidoEm" : {
            "type" : [ "string", "null" ],
            "format" : "date-time"
          },
          "respondidoEm" : {
            "type" : [ "string", "null" ],
            "format" : "date-time"
          },
          "segundosTrabalhadosPrevistos" : {
            "type" : "integer",
            "format" : "int64",
            "description" : "Tempo trabalhado em que o marco dispara",
            "example" : 1800
          },
          "sequencia" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Ordem do marco na categoria, a partir de 1",
            "example" : 1
          },
          "status" : {
            "type" : "string",
            "enum" : [ "AGENDADO", "PENDENTE", "CONCLUIDO", "FALHA", "NAO_ENTREGUE", "NAO_CONCLUIDO", "ADIADO" ]
          },
          "volumeAproximadoMl" : {
            "type" : [ "integer", "null" ],
            "format" : "int32",
            "description" : "Volume para exibir, arredondado para a dezena. Nulo no exercício.",
            "example" : 190
          },
          "volumeMl" : {
            "type" : [ "number", "null" ],
            "description" : "Volume exato: meta ÷ 16. Nulo no exercício.",
            "example" : 187.5
          }
        },
        "required" : [ "bloco", "categoria", "disparadoEm", "editadoEm", "id", "mensagem", "podeAdiar", "podeCorrigir", "recebidoEm", "respondidoEm", "segundosTrabalhadosPrevistos", "sequencia", "status", "volumeAproximadoMl", "volumeMl" ]
      },
...
      "ProblemDetail" : {
        "type" : "object",
        "properties" : {
          "detail" : {
            "type" : "string"
          },
          "instance" : {
            "type" : "string",
            "format" : "uri"
          },
          "properties" : {
            "type" : "object",
            "additionalProperties" : { }
          },
          "status" : {
            "type" : "integer",
            "format" : "int32"
          },
          "title" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string",
            "format" : "uri"
          }
        }
      },
...
    }
  }
}